_site/requirements.txt
pyyaml==6.0.1
requests==2.31.0
numpy==1.26.4